Translate files using LibreTranslate selected instance

With this function, you can translate an external file. Supported formats are .txt, .odt, .odp, .docx, .pptx, .epub and .html

Usage

translate_file(file, from = "auto", to)

Arguments

file

The path to the file you want to translate (absolute or relative).

from

The origin language of the text. If you leave it empty, it will be set to "auto".

to

The destination language you want the text to be translated to.

Examples